% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extrip}
\alias{extrip}
\title{Example itinerary dataset.}
\format{
A data frame with 8 rows of data and 5 variables:
\describe{
  \item{group}{grouping of the itinerary trips, in case of multiple itineraries}
  \item{country}{destination country}
  \item{city}{destination city}
  \item{duration_days}{duration of the stay, in days}
  \item{arrival}{arrival date in YYYY-MM-DD format}
}
}
\usage{
extrip
}
\description{
A dataset containing the itinerary information that can be used as a template.
}
\keyword{datasets}
